pyproject.toml
[project]
name = "br-economic-pulse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "boto3>=1.42.73",
    "duckdb>=1.5.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sentry-sdk>=2.55.0",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
